requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
soupsieve==1.8
urllib3==1.24.1
whitenoise==4.1.2
